Simple & Default view have their own window behaviours now.
In simple view you can't resize the window, and it resizes to the correct height of the radio list. In Detail view you can resize, but only as small as a default size or simple view, whichever is greatest. This takes into account the radio list growing - you won't ever be able to size the window smaller than the radio list (height-wise.)